backend/requirements.txt
Django==4.0.1
psycopg2==2.9.3
pyjwt==2.3.0
bcrypt==3.2.0